# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_mumtaz_reimbursement-system_backend_internal_pkg_apperr.Detail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.ErrorBody'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
  github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_apperr.Detail'
        type: array
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: 'Take-home test, PT Mumtaz Teknologi Indonesia. Cookie-based auth:
    login via /auth/login, then use X-CSRF-Token header on mutations. Session cookies
    are HttpOnly — no bearer tokens.'
  title: Reimbursement Management System API
  version: "1.0"
paths:
...
  /dashboard/category-breakdown:
    get:
      parameters:
      - description: YYYY-MM
        in: query
        name: month
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Per-category totals for a month (?month=YYYY-MM, default current)
      tags:
      - dashboard
  /dashboard/monthly-trend:
    get:
      parameters:
      - description: 1-24
        in: query
        name: months
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Claim totals for the last N months (?months=6, max 24)
      tags:
      - dashboard
...
  /reimbursements:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit (max 100)
        in: query
        name: limit
        type: integer
      - description: Search title/description
        in: query
        name: search
        type: string
      - description: DRAFT | SUBMITTED | APPROVED | REJECTED | PAID | CANCELLED
        in: query
        name: status
        type: string
      - description: Category UUID
        in: query
        name: category_id
        type: string
      - description: YYYY-MM-DD
        in: query
        name: date_from
        type: string
      - description: YYYY-MM-DD
        in: query
        name: date_to
        type: string
      - description: title | amount | status | expense_date | created_at
        in: query
        name: sort
        type: string
      - description: ASC | DESC
        in: query
        name: order
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: 'List claims (scoped: employee=own, manager/finance/admin=all)'
      tags:
      - reimbursements
    post:
      consumes:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Create DRAFT claim with items; amount computed server-side
      tags:
      - reimbursements
  /reimbursements/{id}:
    delete:
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Delete own DRAFT claim → 204
      tags:
      - reimbursements
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Claim detail — items + attachment metadata (scope-checked)
      tags:
      - reimbursements
    patch:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Edit own claim — DRAFT/REJECTED only; items replaced wholesale
      tags:
      - reimbursements
...
  /reimbursements/{id}/submit:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
        "409":
          description: DUPLICATE_SUSPECTED / state conflict
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
        "422":
          description: BUSINESS_RULE_VIOLATED with all violations
          schema:
            $ref: '#/definitions/github_com_mumtaz_reimbursement-system_backend_internal_pkg_response.Envelope'
      summary: Submit claim — full policy check, generates approval steps
      tags:
      - reimbursements
...
securityDefinitions:
  CSRF:
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"